Move hiqp_client into hiqp_ros. Note: Discuss possible namespace changes.
