+++ hiqp_ros/package.xml
@@ -6,7 +6,8 @@
 
   <author email="marcus.adam.johansson@gmail.com">Marcus A Johansson</author>
   <maintainer email="marcus.adam.johansson@gmail.com">Marcus A Johansson</maintainer>
-
+  <maintainer email="ksatyaki@gmail.com">Chittaranjan Srinivas Swaminathan</maintainer>
+  
   <license>GPLv3</license>
 
   <buildtool_depend>catkin</buildtool_depend>
@@ -16,7 +17,6 @@
   <build_depend>controller_interface</build_depend>
   <build_depend>visualization_msgs</build_depend>
   <build_depend>orocos_kdl</build_depend>
-  <!--build_depend>kdl_parser</build_depend-->
   <build_depend>hiqp_core</build_depend>
   <build_depend>hiqp_msgs</build_depend>
 
@@ -24,7 +24,6 @@
   <run_depend>controller_interface</run_depend>
   <run_depend>visualization_msgs</run_depend>
   <run_depend>orocos_kdl</run_depend>
-  <!--run_depend>kdl_parser</run_depend-->
   <run_depend>hiqp_core</run_depend>
   <run_depend>hiqp_msgs</run_depend>
